+++ lib/widgets/app_bar/journal_sliver_appbar.dart
@@ -10,6 +10,7 @@ import 'package:lotti/utils/platform.dart';
 import 'package:lotti/widgets/badges/flagged_badge.dart';
 import 'package:lotti/widgets/badges/tasks_badge_icon.dart';
 import 'package:lotti/widgets/misc/multi_select.dart';
+import 'package:lotti/widgets/misc/tasks_counts.dart';
 import 'package:lotti/widgets/search/search_widget.dart';
 import 'package:lotti/widgets/search/task_status_filter.dart';
 import 'package:lotti/widgets/search/tasks_segmented_control.dart';
@@ -81,9 +82,7 @@ class JournalSliverAppBar extends StatelessWidget {
                                 children: [
                                   Text(
                                     localizations.journalPrivateTooltip,
-                                    style: TextStyle(
-                                      color: styleConfig().secondaryTextColor,
-                                    ),
+                                    style: searchLabelStyle(),
                                   ),
                                   CupertinoSwitch(
                                     value: snapshot.privateEntriesOnly,
@@ -100,9 +99,7 @@ class JournalSliverAppBar extends StatelessWidget {
                               children: [
                                 Text(
                                   localizations.journalFavoriteTooltip,
-                                  style: TextStyle(
-                                    color: styleConfig().secondaryTextColor,
-                                  ),
+                                  style: searchLabelStyle(),
                                 ),
                                 CupertinoSwitch(
                                   value: snapshot.starredEntriesOnly,
@@ -119,9 +116,7 @@ class JournalSliverAppBar extends StatelessWidget {
                                 children: [
                                   Text(
                                     localizations.journalFlaggedTooltip,
-                                    style: TextStyle(
-                                      color: styleConfig().secondaryTextColor,
-                                    ),
+                                    style: searchLabelStyle(),
                                   ),
                                   CupertinoSwitch(
                                     value: snapshot.flaggedEntriesOnly,
@@ -151,6 +146,7 @@ class JournalSliverAppBar extends StatelessWidget {
                             ),
                           ),
                         if (snapshot.showTasks) const TaskStatusFilter(),
+                        if (snapshot.showTasks) const TaskCounts(),
                       ],
                     ),
                   ),